pdf: use original folder structure
for historical reasons (_ahem_) we do not use the given folder
structure when generating pdf but encode the structure into the
file name. this has been done originally in pandoc-lecture for
beamer, but the reasons are lost in history ...
additionally, this approach has some surprising issues since you
cannot use underscores in the filename or path.

let's use the original folder structure like we do for gfm and docsify.

closes #18