shift headings: use h1 as top-level headings to simplify tooling
We currently use H2 as the first header level in all documents.
At the same time, H2 is the slide level, the title is
automatically translated by Pandoc as H1 (well, for some output
formats).

BC, on the other hand, uses H1 to create an organisational
structure  and H2 for the slides. The title of the document is
therefore not automatically used by Pandoc.

Therefore, the YAML variable ‘shift_headings: true’ has been
added in the current tooling for BC's documents, and in the
filters for GFM, Docsify and PDF, the header levels in their
documents are shifted by 1. This is not necessary for the Beamer
output, as otherwise the subheadings wouldn't appear as separate
outline slides.

To simplify the tooling: What would happen if I used H1 as the
top outline level (and slide level) in my documents? We could
then do without the special YAML variable and simplify the Lua
filters (no shifting in there). To make Pandoc set the title as
document title again, we could activate the option  for GFM,
Docsify and PDF output (in the respective default files). It is
still unclear for Beamer, but it should actually also work.

from https://pandoc.org/MANUAL.html#reader-options:
> may be a good choice for converting Markdown documents that
> use level-1 headings for sections to HTML, since pandoc uses
> a level-1 heading to render the document title.

Analysis:

- gfm/docsify (): heading levels shifted (h1 -> h2, h2 -> h3),
  title not inserted as h1 => **shifting necessary**, also we
  need to **add title as h1 manually** for gfm/docsify output
- pdf handout (): heading levels shifted (h1 -> h2, h2 -> h3),
  title not inserted as h1, titlepage using title metadata =>
  **no shifting necessary**, **no extra title as h1 necessary**,
  just use h1 as h1 and title page as is?
- beamer slides (): heading levels shifted (h1 -> h2, h2 -> h3),
  title not inserted as h1, titlepage/slide using title metadata
  => **no shifting necessary**, **no extra title as h1 necessary**,
  just use h1 as h1 and title page as is?

---

**This commit introduces H1 as top-level headings and simplifies
our tooling.**

closes #23