origin: do analyse doc.blocks only
There are certain cases, were we do have something defined in the
YAML header but do not use this in the document or for an output
format - in this case the filter will find this as well and put it
as exception to the license.

Change the behaviour: Do work on doc.blocks only, after the
document structure has been assembled.

fixes #37