gfm/docsify/beamer/pdf: switch to leaner yaml header
instead of defining tldr, outcomes, videos, ... in yaml, let's
put these elements directly into the markdown documents and
wrap them into custom divs. our filter will move this content
to meta data, so in our lua documnt templates we can refer to
meta data. however, there is no need to pre-process this data
anymore, e.g. we do not need to convert the meta data to bullet
points, but use it just as a blocks list and put it into the
document as is. this way the user has more control on how the
special sections should look like.

closes https://github.com/cagix/pandoc-lecture-zen/issues/46