Makefile: test for imagemagick and convert images
if imagemagick is found, then we additionally create a light-version with non-transparent background and suffix '_light' plus a dark-version (inverted colours) with non-transparent background and suffix '_dark'.

otherwise, we just copy all referenced local images as before. so this change should be backwards compatible.

TODO: we need to install imagemagick in our github action.

closes #61